+++ src/artist/artist.controller.ts
@@ -16,8 +16,8 @@ import {
 import { ArtistService } from './artist.service';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
-import { UuidGuard } from '../utils/uuid.guard';
-import { AuthGuard } from 'src/utils/auth.guard';
+import { UuidGuard } from '../guards/uuid.guard';
+import { AuthGuard } from 'src/guards/auth.guard';
 
 @UseGuards(AuthGuard)
 @Controller('artist')
